refactor: deduplicate engine adapters, fix code review items
- Move shared logic (start/stop/health_check/forward_request/stream) to
  EngineAdapter base class. VLLMAdapter: 227→56 lines. SGLangAdapter: 226→57.
  Net -199 lines with zero behavior change.
- Add logging to silent exception handler in health_check()
- Replace semaphore._value private access with tracked counter in TenantManager
- Replace print() with logger.error() in CLI error paths
